requirements-dev.txt
arrow
black
dj-database-url
celery<5.0
flake8==3.7.4
flake8-isort
ipdb
isort==4.3.4
mbq.env>=1,<2
mbq.metrics>=1,<2
psycopg2-binary
pytest!=4.2.0,>=3.6
pytest-django
rollbar
tox
tox-travis
tox-venv
